refactor: add missing ng prefix to the angular module name
replace module name from 'ovh-angular-user-pref' to 'ngOvhUserPref'

BREAKING CHANGE: module is now named as `ngOvhUserPref`
