+++ src/index.js
@@ -1,13 +1,15 @@
 import angular from 'angular';
-import 'ovh-angular-proxy-request';
-import userPrefConstant from './ovh-angular-user-pref.constant';
-import userPrefProvider from './ovh-angular-user-pref.provider';
 
-const moduleName = 'ovh-angular-user-pref';
+import 'ovh-angular-proxy-request';
+
+import constant from './constant';
+import provider from './provider';
+
+const moduleName = 'ngOvhUserPref';
 
 angular
   .module(moduleName, ['ovh-angular-proxy-request'])
-  .constant('OVH_USER_PREF', userPrefConstant)
-  .provider('ovhUserPref', userPrefProvider);
+  .constant('OVH_USER_PREF', constant)
+  .provider('ovhUserPref', provider);
 
 export default moduleName;
